I am so tired of this shit

+++ GameObject.cs
@@ -15,10 +15,11 @@ namespace DonkeyKong
         protected float Speed;
         protected Color Color;
         protected float Rotation;
-        protected float Size;
+        protected int Size;
         protected float LayerDepth;
         protected Vector2 Origin;
-        public GameObject(Texture2D texture, Vector2 position, float speed, Color color, float rotation, float size, float layerDepth, Vector2 origin) 
+
+        public GameObject(Texture2D texture, Vector2 position, float speed, Color color, float rotation, int size, float layerDepth, Vector2 origin) 
         {
             Texture = texture;
             Position = position;
